+++ wp-appkit/lib/user-permissions/auth-engines/rsa-public-private-auth.php
@@ -8,6 +8,10 @@ class WpakRsaPublicPrivateAuth extends WpakAuthEngine {
 
 	public function settings_meta_box_content( $post, $current_box ) {
 
+		//Include theme and addons php files so that authentication hooks are applied
+		WpakThemes::include_app_theme_php( $post->ID );
+		WpakAddons::require_app_addons_php_files( $post->ID );
+
 		$auth_settings = $this->get_authentication_settings( $post->ID );
 
 		$error_message = '';
@@ -59,7 +63,8 @@ class WpakRsaPublicPrivateAuth extends WpakAuthEngine {
 		}
 
 		if ( !empty( $auth_settings['private_key'] ) ) {
-			?><h4><?php _e( 'User connections', WpAppKit::i18n_domain ) ?></h4><?php
+			?><h4><?php _e( 'User connections', WpAppKit::i18n_domain ) ?></h4>
+			<?php
 			$current_connections = $this->get_current_connections( $post->ID );
 			?>
 			<p class="description">
@@ -79,6 +84,8 @@ class WpakRsaPublicPrivateAuth extends WpakAuthEngine {
 						<th style="width:25%"><?php _e( 'Device ID', WpAppKit::i18n_domain ) ?></th>
 						<th style="width:25%"><?php _e( 'Login time', WpAppKit::i18n_domain ) ?></th>
 						<th style="width:25%"><?php _e( 'Last access time', WpAppKit::i18n_domain ) ?></th>
+						<th style="width:25%"><?php _e( 'Validity', WpAppKit::i18n_domain ) ?></th>
+						<th style="width:25%"><?php _e( 'Expiration time', WpAppKit::i18n_domain ) ?></th>
 					</tr>
 				</thead>
 				<tbody>
@@ -108,6 +115,24 @@ class WpakRsaPublicPrivateAuth extends WpakAuthEngine {
 							<?php endforeach; ?>
 							</table>
 						</td>
+						<td>
+							<?php
+								$user_last_time = $this->get_user_last_time( $user_id, $post->ID );
+								if ( $user_last_time ) {
+									$expiration_type = $this->get_expiration_type($user_id, $post->ID);
+									$validity_duration = $this->get_expiration_time($user_id, $post->ID);
+									echo human_time_diff( 0, $validity_duration ) .' ';
+									echo $expiration_type == 'last_access_time' ? __( 'from last access time', WpAppKit::i18n_domain ) : __( 'from login time', WpAppKit::i18n_domain );
+								}
+							?>
+						</td>
+						<td>
+							<?php
+								if ( $user_last_time ) {
+									echo get_date_from_gmt( date( 'Y-m-d H:i:s', $user_last_time + $validity_duration ) );
+								}
+							?>
+						</td>
 					</tr>
 					<?php
 					$cpt++;
@@ -132,7 +157,6 @@ class WpakRsaPublicPrivateAuth extends WpakAuthEngine {
 		$user_meta = '_wpak_auth_'. $app_id;
 		$current_connections = [];
 		$current_connections_raw = $wpdb->get_results( "SELECT user_id, meta_value FROM {$wpdb->prefix}usermeta WHERE meta_key = '$user_meta'", ARRAY_A );
-		//$current_connections_raw
 		if ( !empty( $current_connections_raw ) ) {
 			foreach( $current_connections_raw as $current_connection ) {
 				$user_auth_data = unserialize($current_connection['meta_value']);
@@ -754,10 +778,14 @@ class WpakRsaPublicPrivateAuth extends WpakAuthEngine {
 				return;
 			}
 
-			$purge_time = apply_filters( 'wpak_auth_purge_time', 30*24*3600, $user_id, $app_id ); //1 month by default
-
 			$expiration_type = $this->get_expiration_type( $user_id, $app_id );
 			$expiration_time = $this->get_expiration_time( $user_id, $app_id );
+
+			//Purge time. 1 month by default :
+			$purge_time = apply_filters( 'wpak_auth_purge_time', 30*24*3600, $user_id, $app_id );
+			if ( $purge_time < $expiration_time ) {
+				$purge_time = $expiration_time; //Purge time must be >= to expiration time
+			}
 
 			$time = time();
 			$changed = false;
@@ -768,8 +796,8 @@ class WpakRsaPublicPrivateAuth extends WpakAuthEngine {
 					unset( $user_auth_data[$device_id] );
 					$changed = true;
 				}
+				//Check if all devices have expired
 				if ( $time - $auth_data[$expiration_type] < $expiration_time ) {
-					unset( $user_auth_data[$device_id] );
 					$all_expired = false;
 				}
 			}
@@ -831,6 +859,31 @@ class WpakRsaPublicPrivateAuth extends WpakAuthEngine {
 	protected function get_user_connection_validity( $user_id, $app_id ) {
 		$connection_validity = 0;
 
+		$user_last_time = $this->get_user_last_time( $user_id, $app_id );
+		if ( $user_last_time ) {
+
+			$expiration_time = $this->get_expiration_time( $user_id, $app_id );
+
+			if ( $expiration_time === -1 ) {
+				$connection_validity = 1;
+			} else {
+				$connection_validity = ( ( time() - $user_last_time ) <= $expiration_time ) ? 1 : -1;
+			}
+		}
+
+		return $connection_validity;
+	}
+
+	/**
+	 * Get the given user last login or access time (according to expiration_type)
+	 *
+	 * @param int  $user_id              User id
+	 * @param int  $app_id               App id
+	 * @return int $user_last_time       Timestamp or 0 if no valid last time found
+	 */
+	protected function get_user_last_time( $user_id, $app_id ) {
+		$user_last_time = 0;
+
 		$user_meta = '_wpak_auth_'. $app_id;
 
 		$user_auth_data = get_user_meta( $user_id, $user_meta, true );
@@ -868,22 +921,14 @@ class WpakRsaPublicPrivateAuth extends WpakAuthEngine {
 			$expiration_type = $this->get_expiration_type( $user_id, $app_id );
 
 			if ( $expiration_type === 'login_time' ) {
-				$user_secret_time = $last_login_time;
+				$user_last_time = $last_login_time;
 			} else if ( $expiration_type === 'last_access_time' ) {
-				$user_secret_time = $last_access_time;
-			}
-
-			$expiration_time = $this->get_expiration_time( $user_id, $app_id );
-
-			if ( $expiration_time === -1 ) {
-				$connection_validity = 1;
-			} else {
-				$connection_validity = ( ( time() - $user_secret_time ) <= $expiration_time ) ? 1 : -1;
+				$user_last_time = $last_access_time;
 			}
 
 		}
 
-		return $connection_validity;
+		return $user_last_time;
 	}
 
 	/**